OnSeparate now reports collider parts, and OnSeparate/OnCollide now return void
Reporting collider parts in OnSeparate allows objects to track which parts they
lose contact with for a specific object. An example of where this may be useful
is detecting when separation occurs with a non-solid sensor part versus a solid
body part.

OnSeparate and OnCollide used to return orxBOOL as way to prevent the other
collider's callback from being called. However the callback order can not be
controlled from orx so this simplifies the API by changing the return value
from orxBOOL to void. With this change both colliders, if they are Scroll
objects, will have their OnSeparate and OnCollide callbacks called.